+++ src/app/clusters/push-av-stream-transport-server/PushAVStreamTransportCluster.h
@@ -20,7 +20,7 @@
 
 #include <app/clusters/push-av-stream-transport-server/constants.h>
 #include <app/clusters/push-av-stream-transport-server/push-av-stream-transport-delegate.h>
-#include <app/clusters/push-av-stream-transport-server/push-av-stream-transport-logic.h>
+#include <app/clusters/push-av-stream-transport-server/PushAVStreamTransportLogic.h>
 #include <app/server-cluster/DefaultServerCluster.h>
 
 namespace chip {
