+++ src/app/clusters/push-av-stream-transport-server/app_config_dependent_sources.cmake
@@ -16,10 +16,10 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/push-av-stream-transport-cluster.cpp"
-    "${CLUSTER_DIR}/push-av-stream-transport-cluster.h"
-    "${CLUSTER_DIR}/push-av-stream-transport-logic.cpp"
-    "${CLUSTER_DIR}/push-av-stream-transport-logic.h"
+    "${CLUSTER_DIR}/PushAVStreamTransportCluster.cpp"
+    "${CLUSTER_DIR}/PushAVStreamTransportCluster.h"
+    "${CLUSTER_DIR}/PushAVStreamTransportLogic.cpp"
+    "${CLUSTER_DIR}/PushAVStreamTransportLogic.h"
     "${CLUSTER_DIR}/push-av-stream-transport-delegate.h"
     "${CLUSTER_DIR}/CodegenIntegration.cpp"
     "${CLUSTER_DIR}/CodegenIntegration.h"
